+++ app/models/concerns/has_social_profiles.rb
@@ -4,13 +4,15 @@ module HasSocialProfiles
   PREFIXES = {
     github: "github.com/",
     twitter: "twitter.com/",
-    linkedin: "linkedin.com/in/"
+    linkedin: "linkedin.com/in/",
+    stack_overflow: "stackoverflow.com/users/"
   }
 
   included do
     before_save :normalize_github, if: :will_save_change_to_github?
     before_save :normalize_twitter, if: :will_save_change_to_twitter?
     before_save :normalize_linkedin, if: :will_save_change_to_linkedin?
+    before_save :normalize_stack_overflow, if: :will_save_change_to_stack_overflow?
 
     private
 
@@ -21,6 +23,10 @@ module HasSocialProfiles
         send(handle)&.delete_prefix!("https://#{prefix}")
         send(handle)&.delete_prefix!("http://www.#{prefix}")
         send(handle)&.delete_prefix!("https://www.#{prefix}")
+
+        if handle === :stack_overflow
+          self.stack_overflow = stack_overflow&.split("/")&.first
+        end
       end
     end
   end