+++ test/models/open_startup/reporting_test.rb
@@ -4,7 +4,6 @@ class OpenStartup::ReportingTest < ActiveSupport::TestCase
   include StripeHelper
 
   test "refreshes Revenue records (grouped by month and description)" do
-    skip "Skip until PR #565."
     @balance_transactions = [
       charge(amount: 1000, created: january, description: "New subscription"),
       charge(amount: 2000, created: january, description: "New subscription"),
@@ -24,7 +23,6 @@ class OpenStartup::ReportingTest < ActiveSupport::TestCase
   end
 
   test "refreshes Expense records (grouped by month and merged with additional fees) and adds manual expenses" do
-    skip "Skip until PR #565."
     @balance_transactions = [
       charge(created: january, fee: -100),
       charge(created: january, fee: -200),
@@ -45,7 +43,6 @@ class OpenStartup::ReportingTest < ActiveSupport::TestCase
   end
 
   test "refreshes Contribution records (grouped by month) and adds manual expenses" do
-    skip "Skip until PR #565."
     @balance_transactions = [
       contribution(created: january, amount: 10),
       contribution(created: january, amount: 20),
@@ -62,7 +59,6 @@ class OpenStartup::ReportingTest < ActiveSupport::TestCase
   end
 
   test "refreshes MonthlyBalance records" do
-    skip "Skip until PR #565."
     @balance_transactions = [
       charge(amount: 1000, created: january),
       fee(created: january, amount: -10),
@@ -83,7 +79,6 @@ class OpenStartup::ReportingTest < ActiveSupport::TestCase
   end
 
   test "creates a Metric record for today with MRR and visitors" do
-    skip "Skip until PR #565."
     @subscriptions = [
       Subscription.new(amount: 9900),
       Subscription.new(amount: 9900),