+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 ## 2022
 
+## August
+
+* August 6 - Fixes a bug where responding to messages via email wouldn't send
+
 ### July
 
+* July 30 - Migrate from Action Mailbox to Postmark JSON API - #588
+* July 27 - Change developer bio text - #560 @kaushikhande
+* July 20 - Normalize business website link - #577 @ypcethan
+* July 19 - Add Stack Overflow link to developer profile - #581 @ypcethan
+* July 13 - Update to Pay 4.0.1 and Stripe 6.5.0 - #574
 * July 12 - Update to rails 7.0.3.1
 * July 11 - Invisible captcha when registering for an account - #563
 * July 11 - Fix scroll to bottom and autofocus on conversation page - #521 @pkayokay